+++ CHANGELOG.rst
@@ -9,8 +9,11 @@ Changelog
 * To hopefully fix the ``SSL: CERTIFICATE_VERIFY_FAILED`` errors that
   some users encounter when downloading timescale files, took the risk
   of switching away from Python’s default SSL certificates to the
-  certificate bundle in the `certifi` package.
+  certificate bundle in the ``certifi`` package.
   `#317 <https://github.com/skyfielders/python-skyfield/issues/317>`_
+
+* Added a new almanac routine for finding :ref:`lunar-nodes`.
+  `#361 <https://github.com/skyfielders/python-skyfield/issues/361>`_
 
 * Gave topos objects a new :meth:`~skyfield.toposlib.Topos.itrf_xyz()`
   method that returns their raw ITRF coordinates.